requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
rembg==2.0.53
python-multipart==0.0.9
pillow==10.2.0
